+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## Version 1.2.26
+## Version 1.3.0
 
-  * 
+  * Add fetching of file names for events (unable to download)
+  * Add fetching of payment schedules - Osm::OnlinePayment::Schedule
+  * Add fetching of member's payments for a schedule
+  * Add updating a member's payment status
 
 ## Version 1.2.25